+++ examples/i18n/app/[lang]/layout.tsx
@@ -2,7 +2,6 @@ import 'fumadocs-ui/style.css';
 import { RootProvider } from 'fumadocs-ui/provider';
 import { Inter } from 'next/font/google';
 import type { ReactNode } from 'react';
-import { I18nProvider } from 'fumadocs-ui/i18n';
 
 const inter = Inter({
   subsets: ['latin'],
@@ -25,20 +24,20 @@ export default async function Layout({
           minHeight: '100vh',
         }}
       >
-        <I18nProvider
-          locale={lang}
-          locales={[
-            {
-              name: 'English',
-              locale: 'en',
-            },
-            {
-              name: 'Chinese',
-              locale: 'cn',
-            },
-          ]}
-          translations={
-            {
+        <RootProvider
+          i18n={{
+            locale: lang,
+            locales: [
+              {
+                name: 'English',
+                locale: 'en',
+              },
+              {
+                name: 'Chinese',
+                locale: 'cn',
+              },
+            ],
+            translations: {
               cn: {
                 toc: '目錄',
                 search: '搜尋文檔',
@@ -48,11 +47,11 @@ export default async function Layout({
                 nextPage: '下一頁',
                 chooseLanguage: '選擇語言',
               },
-            }[lang]
-          }
+            }[lang],
+          }}
         >
-          <RootProvider>{children}</RootProvider>
-        </I18nProvider>
+          {children}
+        </RootProvider>
       </body>
     </html>
   );