+++ packages/ui/src/contexts/i18n.tsx
@@ -1,7 +1,5 @@
 'use client';
-import { createContext, usePathname, useRouter } from 'fumadocs-core/framework';
-import { useEffectEvent } from 'fumadocs-core/utils/use-effect-event';
-import { type ReactNode, useMemo } from 'react';
+import { createContext, useContext } from 'react';
 
 export interface Translations {
   search: string;
@@ -43,7 +41,7 @@ export const defaultTranslations: Translations = {
   editOnGithub: 'Edit on GitHub',
 };
 
-export const I18nContext = createContext<I18nContextType>('I18nContext', {
+export const I18nContext = createContext<I18nContextType>({
   text: defaultTranslations,
 });
 
@@ -54,81 +52,5 @@ export function I18nLabel(props: { label: keyof Translations }): string {
 }
 
 export function useI18n(): I18nContextType {
-  return I18nContext.use();
-}
-
-export interface I18nProviderProps {
-  /**
-   * Current locale
-   */
-  locale: string;
-
-  /**
-   * Handle changes to the locale, redirect user when not specified.
-   */
-  onLocaleChange?: (v: string) => void;
-
-  /**
-   * Translations of current locale
-   */
-  translations?: Partial<Translations>;
-
-  /**
-   * Available languages
-   */
-  locales?: LocaleItem[];
-
-  children?: ReactNode;
-}
-
-export function I18nProvider({
-  locales = [],
-  locale,
-  onChange: _onChange,
-  onLocaleChange = _onChange,
-  ...props
-}: I18nProviderProps & {
-  // TODO: remove next major
-  /**
-   * @deprecated use `onLocaleChange` instead
-   */
-  onChange?: I18nProviderProps['onLocaleChange'];
-}) {
-  const router = useRouter();
-  const pathname = usePathname();
-  const onChange = useEffectEvent((value: string) => {
-    if (onLocaleChange) {
-      return onLocaleChange(value);
-    }
-    const segments = pathname.split('/').filter((v) => v.length > 0);
-
-    // If locale prefix hidden
-    if (segments[0] !== locale) {
-      segments.unshift(value);
-    } else {
-      segments[0] = value;
-    }
-
-    router.push(`/${segments.join('/')}`);
-    router.refresh();
-  });
-
-  return (
-    <I18nContext.Provider
-      value={useMemo(
-        () => ({
-          locale,
-          locales,
-          text: {
-            ...defaultTranslations,
-            ...props.translations,
-          },
-          onChange,
-        }),
-        [locale, locales, onChange, props.translations],
-      )}
-    >
-      {props.children}
-    </I18nContext.Provider>
-  );
+  return useContext(I18nContext);
 }